backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0
aiosqlite==0.19.0
deepface==0.0.89
resemblyzer==0.1.3
opencv-python-headless==4.9.0.80
numpy>=1.26.3
scipy>=1.12.0
httpx==0.26.0
fpdf2==2.7.8
python-dotenv==1.0.0
Pillow>=10.2.0
librosa>=0.10.1
soundfile>=0.12.1
PyJWT[crypto]>=2.8.0
